ai-agent/requirements.txt
faiss-cpu
sentence-transformers
pypdf
langchain
langchain-text-splitters
numpy
python-dotenv
groq
